+++ app/cli/ingest.py
@@ -2,11 +2,11 @@ import typer
 from sqlmodel import Session
 
 from app.db.session import engine
-from app.services.embrapa.processing_ingestor import ProcessingIngestor
-from app.services.embrapa.production_ingestor import ProductionIngestor
+from app.production.ingestor import ProductionIngestor
 from app.services.embrapa.commercialization_ingestor import CommercializationIngestor
-from app.services.embrapa.importation_ingestor import ImportationIngestor
 from app.services.embrapa.exportation_ingestor import ExportationIngestor
+from app.services.embrapa.importation_ingestor import ImportationIngestor
+from app.services.embrapa.processing_ingestor import ProcessingIngestor
 
 app = typer.Typer()
